+++ naturtag/app/controls.py
@@ -44,10 +44,10 @@ class Toolbar(QToolBar):
             shortcut='Ctrl+V',
         )
         self.clear_button = self.add_button(
-            '&Clear', tooltip='Clear open images', icon='fa.remove', shortcut='Ctrl+Shift+X'
+            '&Clear', tooltip='Clear open images', icon='fa.remove', shortcut='Ctrl+X'
         )
         self.refresh_button = self.add_button(
-            '&Refresh',
+            '&Refresh tags',
             tooltip='Refresh previously tagged images with latest observation/taxon data',
             icon='fa.refresh',
             shortcut='F5',
@@ -64,11 +64,16 @@ class Toolbar(QToolBar):
         )
 
         # Extra actions not added to the toolbar, but used by the menu
+        self.reset_db_button = self.add_button(
+            '&Reset database',
+            tooltip='Reset local taxonomy & observation database',
+            icon='mdi6.database-remove',
+            visible=False,
+        )
         self.settings_button = self.add_button(
             '&Settings',
             tooltip='Settings',
             icon='fa.gear',
-            shortcut='Ctrl+Shift+S',
             visible=False,
         )
         self.exit_button = self.add_button(
@@ -134,6 +139,9 @@ class Toolbar(QToolBar):
         view_menu = menu.addMenu('&View')
         view_menu.addAction(self.fullscreen_button)
         view_menu.addAction(self.show_toolbar_button)
+
+        tools_menu = menu.addMenu('&Tools')
+        tools_menu.addAction(self.reset_db_button)
 
         settings_menu = menu.addMenu('&Settings')
         settings_menu.addAction(self.settings_button)